+++ src/CrossPointSettings.cpp
@@ -243,6 +243,9 @@ uint8_t migrateTiltDirectionValue(const uint8_t direction) {
 }  // namespace
 
 const char* CrossPointSettings::getDefaultDeviceName() {
+#if (defined(FREEINK_DEVICE_X4CLASSIC) && FREEINK_DEVICE_X4CLASSIC) || defined(SIMULATOR_DEVICE_X4_CLASSIC)
+  return "X4 Classic";
+#endif
   if (BoardConfig::isSticky()) return "Sticky";
   if (BoardConfig::isX4Pro()) return "InkCap X4 Pro";
   if (gpio.deviceIsX3()) return "InkCap X3";